+++ src/test/kotlin/me/bossm0n5t3r/leetcode/binaryTreeLevelOrderTraversal/BinaryTreeLevelOrderTraversalTest.kt
@@ -1,12 +1,12 @@
-package binaryTreeLevelOrderTraversal
+package me.bossm0n5t3r.leetcode.binaryTreeLevelOrderTraversal
 
+import me.bossm0n5t3r.leetcode.utils.TreeNode
+import me.bossm0n5t3r.leetcode.utils.TreeNodeUtil
 import org.junit.jupiter.api.Test
-import utils.TreeNode
-import utils.TreeNodeUtil
 import kotlin.test.assertEquals
 
 class BinaryTreeLevelOrderTraversalTest {
-    private val binaryTreeLevelOrderTraversal = BinaryTreeLevelOrderTraversal()
+    private val binaryTreeLevelOrderTraversal = BinaryTreeLevelOrderTraversal.Solution()
     private val treeNodeUtil = TreeNodeUtil()
 
     data class BinaryTreeLevelOrderTraversalTestData(
